Har tilføjet en ny komponent til at holde på Reservations uge funktionalitet
