+++ src/components/Reservation/ReservationWeek.tsx
@@ -1,0 +1,66 @@
+import { useState } from "react";
+
+declare global {
+    interface Date {
+        getWeek(): number;
+    }
+}
+
+function ReservationWeek() {
+    const [currentWeek, setCurrentWeek] = useState(new Date().getWeek());
+
+    const handlePrevWeek = () => {
+        setCurrentWeek(currentWeek - 1);
+    };
+
+    const handleNextWeek = () => {
+        setCurrentWeek(currentWeek + 1);
+    };
+
+    return (
+        <div className="row row-cols-3">
+            <button className="btn btn-outline-secondary" type="button" onClick={handlePrevWeek}>
+                <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    width="16"
+                    height="16"
+                    fill="currentColor"
+                    className="bi bi-arrow-left"
+                    viewBox="0 0 16 16"
+                >
+                    <path
+                        fillRule="evenodd"
+                        d="M15 8a.5.5 0 0 0-.5-.5H2.707l3.147-3.146a.5.5 0 1 0-.708-.708l-4 4a.5.5 0 0 0 0 .708l4 4a.5.5 0 0 0 .708-.708L2.707 8.5H14.5A.5.5 0 0 0 15 8"
+                    />
+                </svg>
+            </button>
+            <p className="text-center">{`Uge ${currentWeek}`}</p>
+            <button className="btn btn-outline-secondary" type="button" onClick={handleNextWeek}>
+                <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    width="16"
+                    height="16"
+                    fill="currentColor"
+                    className="bi bi-arrow-right"
+                    viewBox="0 0 16 16"
+                >
+                    <path
+                        fillRule="evenodd"
+                        d="M1 8a.5.5 0 0 1 .5-.5h11.793l-3.147-3.146a.5.5 0 0 1 .708-.708l4 4a.5.5 0 0 1 0 .708l-4 4a.5.5 0 0 1-.708-.708L13.293 8.5H1.5A.5.5 0 0 1 1 8"
+                    />
+                </svg>
+            </button>
+        </div>
+    );
+}
+
+// Extend the Date object with a method to get the week number
+Date.prototype.getWeek = function (): number {
+    const d: Date = new Date(Date.UTC(this.getFullYear(), this.getMonth(), this.getDate()));
+    const dayNum: number = d.getUTCDay() || 7;
+    d.setUTCDate(d.getUTCDate() + 4 - dayNum);
+    const yearStart: Date = new Date(Date.UTC(d.getUTCFullYear(), 0, 1));
+    return Math.ceil(((d.getTime() - yearStart.getTime()) / 86400000 + 1) / 7);
+};
+
+export default ReservationWeek;
